Refactor DB class to src/Utils/. DB class is used to validate collation of live DB tables and convert them to same collation as local tables, not involved in any migration, and is more accurately a util.
